Add spawn host card to table
